Incorporate review suggestions
